config,sqlbase: move computation of splits for SQL tables to sqlbase
Before this patch, the config package was in charge of figuring out the
split points to be created by the split queue - generally between tables
and between partitions where there's an applicable zone config.
This patch moves the code for determining the splits to sqlbase, where
it belongs better: there was no good reason for the config package to
deal with table keys and, in the future we'll likely tie zone configs
and table descriptors more and more (and the config package doesn't have
access to table descriptor).

This patch also rationalizes the remaining code dealing with splits in
the config package, which was flirting with unreadability.

After this patch (as before it), static splits below the user key space
(i.e. /Table/50) are determined by the config package
(SystemConfig.ComputeSplitKey()). To keep the code simple, that method
now asserts that none of the objects in the system key space have zone
configs with subzones; where there such configs, the code would not be
equipped to deal with creating splits for these subzones. The relevant
code has moved to sqlbase, and is only invoked for user tables.

Release note: None