roachprod: plumb program-filter flag
Before this PR we'd always use the default cockroach program filter which is
`^cockroach.*`, the problem with this filter is that now it picks up the very
verbose rocksdb logs which live in files like:

```
cockroach-rocksdb.ip-10-12-37-172.ubuntu.2019-10-17T15_31_15Z.004090.log
```

Which have an effective program name of `cockroach-rocksdb`.

It might be reasonable to default this value to the value of `--binary`
flag but that would lead to weird usability issues if the remote cockroach
is called something different from the local one.

Release note: None