storage/engine: don't set iterator bounds in pebbleMVCCScanner
The RocksDB `mvccScanner` doesn't set the iterator bounds, relying
instead on the caller to do so. I believe the `pebbleMVCCScanner` was
doing this to make `MVCCGet` work correctly, but it seems the problem
there was that we weren't using prefix iteration for
`pebbleMVCCScanner.get()`. Fixing that allows us to elide the bounds
setting.

Release note: None