opt: clear column stats from prepared memo
This commit clears all column statistics from every relational expression in
the memo when the memo is created as part of a prepared statement. The purpose
of this change is to free up the potentially large amount of memory used by
histograms.

This does not affect the quality of the plan used at execution time, since
the stats are just recalculated anyway when placeholders are assigned.

Fixes #41524

Release note: None