bulk: sstBatcher default to disallowShadowing
IMPORT's batcher is configured by its BufferingAdder to set this to true, but RESTORE uses the raw batcher.
When false, mvcc stats are marked as estimates as the SST might, or might not, be adding the keys it contains.
When true, we *know* every key in the SST is new (or it would be an error), so we can mark stats as exact.

Defaulting to true will mean RESTORE's SSTables are sent with DisallowShadowing=true, meaning their stats
should be marked as exact. They check should be very cheap in RESTORE as it should take the empty-span fast-path,
and having stats not marked as estimates is cheaper in the long-term (avoids recomputations later, e.g. in splits).

Release note: none.