colexec: fix hash joiner when the build table is empty
Previously, with an outer hash join with an empty build table but non-empty
probe table we would hit index out of bounds error. The root problem is
that we always copy the values vector for the build table regardless
of whether the corresponding rows actually matched. Now before copying
we make sure that the hash table is not empty and also for all unmatched
rows we set buildIdx to 0 so that we can copy the garbage zeroth row
(the setting of NULLs later will take care of the correctness).

Release note: None