multiregionccl: introduce data-driven test framework for multi-region
This patch introduces a new data-driven test framework for multi-region
syntax. TestMultiRegionDataDriven contains a full description of how
to use this thing.

There were a few motivations to introduce this. Firstly, this makes it
very easy to create specific cluster topologies, which is quite useful
when testing multi-region invariants at a high level. Secondly, this
allows us to create directives that analyze and assert invariants in
query traces easily. This also allows us to create directives to wait
for zone config updates to take place which is quite neat.

I've also added a test for global tables that uses this framework. The
test ensures that once new regions are added to the database, present
time reads can be served via (local) follower reads from the new
region.

Release note: None