concurrency: elide signaling new state channel on some no-op updates
Previously, various places in the lock table code
(e.g. `informActiveWaiters`) would update a request's waiting state,
and in a few cases, unconditionally signal the new state channel.
This would happen even if the waiting state for the active waiter
had not changed. This was fine, as the contract for signaling the new
state channel makes no guarantees that there has indeed been a state
change -- just that there could have been one.

This patch tightens the cases where the channel is signaled/waiting
state is updated. We elide updates to the waiting state and signaling
the channel in cases where the waiter does not need to take any
meaningful action. This change will allow us to blindly update waiting
states in the future, by calling `informActiveWaiters`, without needing
to decide whether there is a state change or not at the caller.

As a result of this patch, some observability related fields like the
number of waiting readers/writers may be stale. Note that these were
could always be stale, this patch just increases the cases where they
may be stale.

Release note: None